pyproject.toml
[project]
name = "ddframework"
description = "ddframework"
version = "0.0.1"
readme = "README.md"

authors = [
    { name="Michael Lamertz", email="michael.lamertz@gmail.com" }
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Topic :: Games/Entertainment",
    "Topic :: Software Development :: Libraries :: pygame",
]

dependencies = [
    "pygame-ce",
    "pgcooldown",
]

[project.scripts]
sampleapp = "sampleapp.main:main"

[project.urls]
homepage = "https://github.com/dickerdackel/ddframework"
bugtracker = "https://github.com/DickerDackel/ddframework/issues"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
addopts = "-rA -s --cov-report term-missing -s"
testpaths = [
    "tests",
]

[tool.ruff]
ignore = [
    # "ANN",
    "E501",
    "E701",
    "E731",
    "E741",
    "I001",
    "PLR2004",
]
line-length = 100

[tool.ruff.lint.isort]
force-sort-within-sections = false
known-third-party = ["pygame", "numpy", "tinyecs", "patternegine", "pgcooldown", "rpeasings"]
lines-between-types = 1
